Add .cake extension to supported debug file types
Currently, trying to debug `Cake.exe` while executing `.cake` script files will not hit any breakpoints set inside of VS Code since the file type is not supported.

Adding the file extension `.cake` to this list of supported extensions allows us to launch `Cake.exe` with the mono soft debugger, and subsequently attach to it with this extension to debug our cake scripts.