Use mono source for debugger-libs
I was attempting to use this extension to debug a Xamarin.Android
application. It "sort of" works, except the locals render in an odd
way, such as:

    Local
        arg0: 0xfffffffff0b76140
        arg1: 0xffffffffffff894c
        arg2: 0x0

I think the issue is just that the Mono debugger libraries are old and
out of date.

The NuGet packages being used for the Mono debugger are unlisted and
unmaintained, such as:

https://www.nuget.org/packages/Mono.Debugging/1.0.20161020.46

Within Xamarin, most projects add a `git submodule` for
mono/debugger-libs and xamarin/nrefactory and build these libraries
from source. This allowed me to remove NuGet packages and use
`@(ProjectReference)` instead.

I also had to update `$(TargetFrameworkVersion)` to `v4.7.2`.

After doing this, I updated some other NuGet packages so their versions
matched what `debugger-libs` is using:

* Mono.Cecil 0.10.1
* Newtonsoft.Json 10.0.3

Now I can see proper locals in a Xamarin.Android app!

Using a `launch.json`:

    {
        "version": "0.2.0",
        "configurations": [
            {
                "name": "Debug",
                "type": "mono",
                "request": "attach",
                "address": "localhost",
                "port": 10000,
                "preLaunchTask": "Debug",
            }
        ]
    }

And a `tasks.json`:

    {
        "version": "2.0.0",
        "tasks": [
            {
                "label": "Debug",
                "type": "shell",
                "command": "msbuild HelloAndroid/HelloAndroid.csproj -t:Install,_Run -p:AndroidAttachDebugger=true"
            },
        ]
    }

I will setup a .NET 6 example in this repo down the road:

https://github.com/xamarin/net6-samples